### 2025-02-12 #7775 @JCAurre
Hello! Regarding suggested reviewers, we thought of
(1) Gabriele Bozzola (Caltech): He also reviewed GRChombo and GRDzhadzha.
(2) Nicolas Sanchis-Gual (Valencia): He does NR with fundamental fields using a different code.
(3) Harry Shum (Nottingham): He does NR and has been using the solver.
(4) Alessio Notari (Barcelona), Jorge Norena (Valparaiso U) or Yanina Lopez (Valparaiso): They do cosmology and were interested in using the solver.
