### 2025-02-12 #7775 @warrickball
Thanks for your submission, @JCAurre, and for the suggested reviewers. That will serve as a useful starting point for whichever editor picks this up. That said, all the suitable JOSS editors are currently working at capacity so I'm going to "waitlist" this review until an editor with the relevant expertise is available to take it on. Thanks for your patience!

In the meantime, could you comment a bit further on the development history of GRTresna? My impression is that the GitHub repo was created for the JOSS submission but that the code development goes back beyond Dec 4, 2024. I don't doubt that GRTresna represents [substantial scholarly effort](https://joss.readthedocs.io/en/latest/submitting.html#substantial-scholarly-effort) but it would be good to know how much further back the development goes, and if that history can be found somewhere else.

The paper is also currently too long for JOSS. [Our docs say](https://joss.readthedocs.io/en/latest/paper.html#what-should-my-paper-contain)

> The paper should be between 250-1000 words. Authors submitting papers significantly longer than 1000 words may be asked to reduce the length of their paper.

The main target would be to shorten the "Key features" section, much of which can be expanded on in the project documentation. We don't have to get all the way down to 1000 words specifically but the paper should at least be somewhat less than 1500 words.
