### 2025-02-24 30b069c
Include tagging criteria in GRSolver template list and specify in example main functions
