### 2025-02-24 f3b4197
Include tagging directly in test
