### 2025-04-16 8eab52b
if statement to avoid dividing angular momentum over black hole mass if this is zero
