### 2025-08-31 #7775 @Sbozzolo
@rashti-alireza, this code deals with the initial data problem in GR and seems up your ally (or at least, your former ally...). Would you be interest available to review it?
