### 2025-09-14 #7775 @Sbozzolo
@leowerneck has kindly agreed to review this code. Thank you!

I am still working on the second reviewer.
