### 2025-09-22 #9057 @editorialbot
[REVIEW]: GRTresna: An open-source code to solve the initial data constraints in numerical relativity
**Submitting author:** @JCAurre (<a href="http://orcid.org/0000-0001-9584-5791">Josu C. Aurrekoetxea</a>)
**Repository:** https://github.com/GRTLCollaboration/GRTresna
**Branch with paper.md** (empty if default branch):
**Version:** v1.0.0
