### 2026-01-19 #9057 @gabrielepalloni98
I have just finished the review. I am sharing this [Review_GRTresna.pdf](https://github.com/user-attachments/files/24712980/Review_GRTresna.pdf) with you, where I have prepared a brief commentary on the code. In this message, I have included the main comments related to some issues we encountered when compiling the code on our server.

The installation instructions are clearly written. During our review, we raised a minor comment to the authors regarding the **Compiling GRTresna** section, which was promptly addressed and corrected. In particular, since GRTresna requires `C++14`, the Chombo library must be compiled with a compatible standard.

To ensure consistency, depending on the selected compiler, the following changes are required.

In `lib/mk/Make.defs.local,` add the flag `CXXSTD = 14` (or `17`). By default, `lib/mk/compiler/Make.defs.GNU` sets `C++11,` as shown below:

```
ifeq ($(CXXSTD),11)
  _cxxbaseflags := -std=c++11
else ifeq ($(CXXSTD),14)
  _cxxbaseflags := -std=c++14
else ifeq ($(CXXSTD),17)
  _cxxbaseflags := -std=c++17
else
  _cxxbaseflags := -std=c++11
endif
```

In `ib/mk/compiler/Make.defs.clang++`, replace:

```
defcxxdbgflags = -std=c++11 -g -Wall -Wno-overloaded-virtual -Wno-sign-compare
defcxxoptflags = -std=c++11 -O3 -ffast-math -march=native -Rpass=loop-vectorize -Wno-sign-compare
```

with:

```
defcxxdbgflags = -std=c++14 -g -Wall -Wno-overloaded-virtual -Wno-sign-compare
defcxxoptflags = -std=c++14 -O3 -ffast-math -march=native -Rpass=loop-vectorize -Wno-sign-compare
```

In `lib/mk/compiler/Make.defs.Intel`, modify:
```
_cxxbaseflags = -restrict -m64 -std=c++11
```
to:
```
_cxxbaseflags = -restrict -m64 -std=c++14
```

After applying these changes, GRTresna compiles correctly. We have already highlighted these issues to the authors (see Issue [#25](https://github.com/openjournals/joss-reviews/issues/25)).

The smooth and effective interaction with the authors demonstrates how the project fulfills the _Community Guidelines_: users can report issues, contribute to the code, and seek support.

I ran the code working with the examples already written by the authors, using their default parameters (even though in the GitHub Wiki the authors described the parameters, guiding the potential user to modify the examples): the code achieves a convergence that minimizes computational cost, as stated in the code (in particular, using 4 MPI processors, the code can find a solution within 1 day).
