### 2026-02-14 22e535c
Update paper.md to address referees comments
