+++ paper/paper.md
@@ -87,13 +87,13 @@ In the ADM form of the Einstein Equations [@Arnowitt:1962hi], we slice the space
 
 $$ds^2 = -(\alpha^2 - \beta_i\beta^i) dt^2 + 2\beta_i dx^i dt + \gamma_{ij} dx^i dx^j$$
 
-and the elliptic constraints are expressed as
+where $\alpha$ and $\beta^i$ are the lapse and shift functions, respectively. The Hamiltonian and momentum constraints are expressed as
 
 $$H \equiv R + K^2-K_{ij}K^{ij}-16\pi \rho = 0,$$
 
 $$M_i \equiv D^j (K_{ij}- \gamma_{ij} K) - 8\pi S_i = 0.$$
 
-Here, $\gamma_{ij}$ is the 3-metric of the hypersurface, $R$ is the Ricci scalar associated to this metric, and $K_{ij}\sim \partial_t \gamma_{ij}$ is the extrinsic curvature tensor, with $K=\gamma^{ij}K_{ij}$ its trace. The decomposed components of the stress-energy tensor of matter (measured by normal observers) are defined as $\rho = n_\mu n_\nu T^{\mu\nu}$ and $S_i = -\gamma_{i\mu} n_\nu T^{\mu\nu}$, where $n_\mu = (-\alpha,0,0,0)$. These equations constitute the set of four PDEs to be solved. There are 16 unknowns: 6 in $\gamma_{ij}$, 6 in $K_{ij}$, 1 in $\rho$ and 3 in $S_i$. Usually the matter configuration is set by the physical scenario, which determines $\rho$ and $S_i$. The constraints only determine 4 quantities, and the remaining 8 (4 of which are physical degrees of freedom, and 4 gauge choices) must be chosen according to physical principles or knowledge about the system. In this short paper, we introduce GRTresna, an open-source code to solve these equations.
+Here, $\gamma_{ij}$ is the 3-metric of the hypersurface, $R$ is the Ricci scalar associated to this metric, and $K_{ij}\sim \partial_t \gamma_{ij}$ is the extrinsic curvature tensor, with $K=\gamma^{ij}K_{ij}$ its trace. The decomposed components of the stress-energy tensor of matter $T^{\mu\nu}$ (measured by normal observers) are defined as $\rho = n_\mu n_\nu T^{\mu\nu}$ and $S_i = -\gamma_{i\mu} n_\nu T^{\mu\nu}$, where $n_\mu = (-\alpha,0,0,0)$. These equations constitute the set of four PDEs to be solved. There are 16 unknowns: 6 in $\gamma_{ij}$, 6 in $K_{ij}$, 1 in $\rho$ and 3 in $S_i$. Usually the matter configuration is set by the physical scenario, which determines $\rho$ and $S_i$. The constraints only determine 4 quantities, and the remaining 8 (4 of which are physical degrees of freedom, and 4 gauge choices) must be chosen according to physical principles or knowledge about the system. In this short paper, we introduce GRTresna, an open-source code to solve these equations.
 
 
[…]
