### 2026-02-14 7c905e8
Update paper.bib to add citations
