+++ paper/paper.bib
@@ -1,3 +1,28 @@
+@article{Tootle:2025ikk,
+    author = "Tootle, Samuel D. and Werneck, Leonardo R. and Assump{\c{c}}{\~a}o, Thiago and Jacques, Terrence Pierre and Etienne, Zachariah B.",
+    title = "{Accelerating numerical relativity with code generation: CUDA-enabled hyperbolic relaxation}",
+    eprint = "2501.14030",
+    archivePrefix = "arXiv",
+    primaryClass = "gr-qc",
+    doi = "10.1088/1361-6382/add63e",
+    journal = "Class. Quant. Grav.",
+    volume = "42",
+    number = "11",
+    pages = "115005",
+    year = "2025"
+}
+
+@software{deppe_2023_8196313,
+  author    = {Deppe, Nils and Throwe, William and Kidder, Lawrence E. and Vu, Nils L. and Hébert, François and Moxon, Jordan and Nelli, Kyle C. and Armaza, Cristóbal and Bonilla, Marceline S. and Kim, Yoonsoo and Kumar, Prayush and Lovelace, Geoffrey and Macedo, Alexandra and O'Shea, Eamonn and Pfeiffer, Harald P. and Scheel, Mark A. and Teukolsky, Saul A. and Wittek, Nikolas A. and Anantpurkar, Isha and Boyle, Michael and Carpenter, Alexander and Ceja, Andrea and Chaudhary, Himanshu and Foucart, Francois and Ghadiri, Noora and Giesler, Matthew and Guo, Jason S. and Iozzo, Dante A. B. and Jones, Ken Z. and Lara, Guillermo and Legred, Isaac and Li, Dongjun and Ma, Sizheng and Melchor, Denyz and Morales, Marlo and Most, Elias R. and Nee, Peter James and Osorio, Alejandro and Pajkos, Michael A. and Pannone, Kyle and Ramirez, Teresita and Ring, Noah and Rüter, Hannes R. and Sanchez, Jennifer and Stein, Leo C. and Tellez, Daniel and Thomas, Sierra and Vieira, Daniel and Wlodarczyk, Tom and Wu, David},
+  title     = {SpECTRE},
+  month     = jul,
+  year      = {2023},
+  version   = {2023.07.29},
+  publisher = {Zenodo},
+  doi       = {10.5281/zenodo.8196313},
+  url       = {https://doi.org/10.5281/zenodo.8196313}
+}
+
 @article{Vu:2021coj,
     author = "Vu, Nils L. and others",
     title = "{A scalable elliptic solver with task-based parallelism for the SpECTRE numerical relativity code}",
