### 2026-02-17 #9057 @JCAurre
Dear Editor and Referees,

Thank you very much for your hard work and for the helpful feedback on GRTresna. We have now merged your suggested changes into the main branch of the repository ([see here](https://github.com/GRTLCollaboration/GRTresna/pull/26))

Regarding the comments about compiling Chombo with the `Make.defs.local` files, we would be happy to add these to our public repository ([here](https://github.com/GRTLCollaboration/GRChombo/tree/main/InstallNotes/MakeDefsLocalExamples)). However, on systems such as macOS it is quite common for such configuration files to stop working after major Apple updates. We will nevertheless do our best to keep them as up to date as possible.
