### 2026-02-20 #9057 @JCAurre
Hi @Sbozzolo, thank you for generating the pdf! I just realised there seems to be an issue with three citations in line 120.
