### 2026-03-04 #9057 @Sbozzolo
@JCAurre, everything is looking good and we can move to the final steps of the process. I opened two issues to fix some small problems with the paper. In addition to that, could you please check the DOIs, and fix them, if possible. Some pieces of software might be best cited with their paper.

Finally, the paper is on the longer side (in terms of word count). It would be great if you could reduce it by 20 % or so.

Thank you!
