### 2026-03-12 2d90308
Update paper.md to make it shorter
