+++ paper/paper.md
@@ -68,10 +68,8 @@ affiliations:
   index: 7
 - name: Information Services, University of Cambridge, Roger Needham Building, 7 JJ Thomson Avenue, Cambridge, CB3 0WA, United Kingdom
   index: 8
-- name: Institute for Advanced Computational Science, Stony Brook University, NY 11794 USA
-  index: 9
 - name: Department of Physics and Astronomy, Johns Hopkins University, Baltimore, MD 21218, USA
-  index: 10
+  index: 9
 date: 22 Jan 2025
 bibliography: paper.bib
 
@@ -117,15 +115,7 @@ The key features of GRTresna are as follows
 
 - Compatibility: As GRTresna is developed on top of Chombo, the solver is primarily designed to be compatible with GRChombo [@Andrade:2021rbd;@Clough:2015sqa] and the family of codes developed by the GRTL Collaboration. We provide two examples that integrate directly with existing examples in the GRChombo evolution code (via the output of a checkpoint file for restart at $t=0$), and provide guidance and tools to validate the results. However, the code outputs data in the standard HDF5 data format, which should be straightforward to adapt to other NR codes that support HDF5 input or can be accessed using Python.
 
-Other features that are inherited from Chombo include
-
-- C++ class structure: GRTresna is written in the C++ language, and makes heavy use of object-oriented programming (OOP) and templating.
-    
-- Parallelism: GRTresna uses hybrid OpenMP/MPI parallelism.
-    
-- Adaptive Mesh Refinement: The code inherits the flexible AMR grid structure of Chombo, with block-structured Berger-Rigoutsos grid generation [@Berger:1991]. The tagging of refinement regions is fully flexible and while it is based on the sources of the elliptic equations by default, other user-defined measures can be defined [@Radia:2021smk].
-
-- Fast: The code uses a multigrid method to efficiently reduce errors across a hierarchy of discretizations, enabling the solver to achieve rapid convergence while minimizing computational costs. This makes GRTresna highly optimized for handling the demanding computations of initial data in the presence of AMR.
+The code achieves rapid convergence by using a multigrid method inherited from Chombo. Other features that are inherited from Chombo include C++ class structure, hybrid OpenMP/MPI parallelism, Adaptive Mesh Refinement [@Berger:1991;@Radia:2021smk]. See [@Andrade:2021rbd] for more details.
 
 Forthcoming features currently under development include the addition of other methods, in particular the Extended Conformal Thin Sandwich (XCTS) method, non-conformally flat metric data, new matter types including vector fields, the modified scalar-tensor gravity formalism of Brady et. al. [@Brady:2023dgu] and dimensional reduction to 2D using the cartoon formalism [@Alcubierre:1999ab;@Cook:2016soy].
 
@@ -133,7 +123,7 @@ Forthcoming features currently under development include the addition of other m
 # Statement of Need
 
 
[…]
