+++ paper/paper.md
@@ -123,7 +123,7 @@ Forthcoming features currently under development include the addition of other m
 # Statement of Need
 
 
-There are a number of existing initial data solvers for numerical relativity, most of which are primarily designed to solve for initial conditions in compact object mergers (i.e. neutron stars and black holes). These include TwoPunctures [@Ansorg:2004ds], SGRID [@Tichy:2009yr], BAM [@Bruegmann:2006ulg], LORENE [@LORENE;@Gourgoulhon:2000nn], Spells [@Pfeiffer:2002wt], SpECTRE [@Vu:2021coj;@Vu:2024cgf;deppe_2023_8196313] ([@Nee:2024bur] for modified gravity) COCAL [@Uryu:2011ky;@Tsokaros:2012kp;@Tsokaros:2015fea], PCOCAL [@Boukas:2023ckb], Elliptica [@Rashti:2021ihv], NRPyElliptic [@Assumpcao:2021fhq,@Tootle:2025ikk], KADATH/FUKA [@FUKA;@Grandclement:2009ju;@Papenfort:2021hod}], SPHINCS\_ID [@SPHINCSID;@Diener:2022hui;@Rosswog:2023nnl], and the solver of East *et al.* [@East:2012zn]. Many of these codes, particularly those using spectral methods like TwoPunctures and SpECTRE, provide a higher accuracy in the solution compared to GRTresna, which is limited to second order accuracy by the multigrid method used. They are therefore better suited to initial data for waveform generation where precision is key. GRTresna is, however, designed to be more flexible and general purpose, tackling both cosmological and black hole spacetimes in a range of scenarios beyond GR and the Standard Model.
+There are a number of existing initial data solvers for numerical relativity, most of which are primarily designed to solve for initial conditions in compact object mergers (i.e. neutron stars and black holes). These include TwoPunctures [@Ansorg:2004ds], SGRID [@Tichy:2009yr], BAM [@Bruegmann:2006ulg], LORENE [@LORENE;@Gourgoulhon:2000nn], Spells [@Pfeiffer:2002wt], SpECTRE [@Vu:2021coj;@Vu:2024cgf;@deppe_2023_8196313] ([@Nee:2024bur] for modified gravity) COCAL [@Uryu:2011ky;@Tsokaros:2012kp;@Tsokaros:2015fea], PCOCAL [@Boukas:2023ckb], Elliptica [@Rashti:2021ihv], NRPyElliptic [@Assumpcao:2021fhq,@Tootle:2025ikk], KADATH/FUKA [@FUKA;@Grandclement:2009ju;@Papenfort:2021hod}], SPHINCS\_ID [@SPHINCSID;@Diener:2022hui;@Rosswog:2023nnl], and the solver of East *et al.* [@East:2012zn]. Many of these codes, particularly those using spectral methods like TwoPunctures and SpECTRE, provide a higher accuracy in the solution compared to GRTresna, which is limited to second order accuracy by the multigrid method used. They are therefore better suited to initial data for waveform generation where precision is key. GRTresna is, however, designed to be more flexible and general purpose, tackling both cosmological and black hole spacetimes in a range of scenarios beyond GR and the Standard Model.
 
[…]
