### 2026-03-12 3deb927
Update paper.md for JOSS review comments
* Update paper.md to make it shorter

* fix GRChombo to GRTresna in key projects

* added DOI for Cosmology using numerical relativity article

* fix citations

* remove website citations
