+++ paper/paper.bib
@@ -78,14 +78,18 @@
     year = "2012"
 }
 
-@article{Aurrekoetxea:2024mdy,
+@article{Aurrekoetxea:2024ypv,
     author = "Aurrekoetxea, Josu C. and Clough, Katy and Lim, Eugene A.",
     title = "{Cosmology using numerical relativity}",
     eprint = "2409.01939",
     archivePrefix = "arXiv",
     primaryClass = "gr-qc",
-    journal = "",
-    year = "2024"
+    doi = "10.1007/s41114-025-00058-z",
+    journal = "Living Rev. Rel.",
+    volume = "28",
+    number = "1",
+    pages = "5",
+    year = "2025"
 }
 
 @article{Aurrekoetxea:2023fhl,
