+++ paper/paper.md
@@ -75,7 +75,7 @@ bibliography: paper.bib
 
 ---
 
-GRTresna is a multigrid solver designed to solve the constraint equations for the initial data required in numerical relativity simulations. In particular it is focussed on scenarios with fundamental fields around black holes and inhomogeneous cosmological spacetimes. The following overview has been prepared as part of the submission of the code to the Journal of Open Source Software. The code is based on the formalism in Aurrekoetxea, Clough \& Lim [@Aurrekoetxea:2022mpw] and can be found at https://github.com/GRTLCollaboration/GRTresna
+GRTresna is a multigrid solver designed to solve the constraint equations for the initial data required in numerical relativity simulations. In particular it is focussed on scenarios with fundamental fields around black holes and inhomogeneous cosmological spacetimes. The following overview has been prepared as part of the submission of the code to the Journal of Open Source Software. The code is based on the formalism in Aurrekoetxea, Clough \& Lim [@Aurrekoetxea:2022mpw] and can be found at [github.com/GRTLCollaboration/GRTresna](github.com/GRTLCollaboration/GRTresna)
 
 # Summary
 
@@ -91,10 +91,10 @@ $$H \equiv R + K^2-K_{ij}K^{ij}-16\pi \rho = 0,$$
 
 $$M_i \equiv D^j (K_{ij}- \gamma_{ij} K) - 8\pi S_i = 0.$$
 
-Here, $\gamma_{ij}$ is the 3-metric of the hypersurface, $R$ is the Ricci scalar associated to this metric, and $K_{ij}\sim \partial_t \gamma_{ij}$ is the extrinsic curvature tensor, with $K=\gamma^{ij}K_{ij}$ its trace. The decomposed components of the stress-energy tensor of matter $T^{\mu\nu}$ (measured by normal observers) are defined as $\rho = n_\mu n_\nu T^{\mu\nu}$ and $S_i = -\gamma_{i\mu} n_\nu T^{\mu\nu}$, where $n_\mu = (-\alpha,0,0,0)$. These equations constitute the set of four PDEs to be solved. There are 16 unknowns: 6 in $\gamma_{ij}$, 6 in $K_{ij}$, 1 in $\rho$ and 3 in $S_i$. Usually the matter configuration is set by the physical scenario, which determines $\rho$ and $S_i$. The constraints only determine 4 quantities, and the remaining 8 (4 of which are physical degrees of freedom, and 4 gauge choices) must be chosen according to physical principles or knowledge about the system. In this short paper, we introduce GRTresna, an open-source code to solve these equations.
+Here, $\gamma_{ij}$ is the 3-metric of the hypersurface, $R$ is the Ricci scalar associated to this metric, and $K_{ij}\sim \partial_t \gamma_{ij}$ is the extrinsic curvature tensor, with $K=\gamma^{ij}K_{ij}$ its trace. The decomposed components of the stress-energy tensor of matter $T^{\mu\nu}$ (measured by normal observers) are defined as $\rho = n_\mu n_\nu T^{\mu\nu}$ and $S_i = -\gamma_{i\mu} n_\nu T^{\mu\nu}$, where $n_\mu = (-\alpha,0,0,0)$. These equations constitute the set of four PDEs to be solved. In this short paper, we introduce GRTresna, an open-source code to solve these equations.
 
 
[…]
