### 2026-03-13 f02b1a2
Update paper.md
