### 2026-03-31 #9057 @Sbozzolo
@JCAurre, thank you!

The author list on zenodo has a slightly different order than the one in the paper. (Author 3 and 4 are swapped). It would be best if the two matched.

Moreover, Shibata's book is missing some metadata (e.g., publisher, see Baumgarte's book)
