### 2026-03-31 #9057 @Sbozzolo
I also noticed that the spelling of the names of one of the authors does not match the ORCID entry. I opened a PR to fix that, but feel free to close if the author is fine with the spelling.
https://github.com/GRTLCollaboration/GRTresna/pull/35
